renamed layout files
