+++ client/src/pages/store/checkout/Success.js
@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import { NavLink } from "react-router-dom";
 
 import useCart from "hooks/useCart";
-import LayoutStoreHome from "layouts/LayoutStoreHome";
+import LayoutStoreHome from "layouts/StoreHome";
 
 const StoreSuccess = () => {
   const { clearCartItems } = useCart();
